Move effects tests to contrib/
