+++ crates/oak_semantic/tests/integration/contrib.rs
@@ -1,0 +1,6 @@
+mod base;
+mod magrittr;
+mod rlang;
+mod s7;
+mod shiny;
+mod testthat;
