+++ crates/oak_semantic/tests/integration/main.rs
@@ -1,4 +1,5 @@
 mod builder;
-mod builder_nse;
+mod common;
+mod contrib;
 mod resolvers;
 mod use_def_map;
